(refactor) Refactored the middleware functions
